HIDE UI ver/server/requirements.txt
amqp==2.5.0
anyjson==0.3.3
asgiref==3.2.7
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.41.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-filter==2.2.0
django-rest-auth==0.9.5
django-timezone-field==3.0
djangorestframework==3.9.4
idna==2.8
importlib-metadata==1.6.0
kombu==4.6.3
Markdown==3.2.2
numpy==1.16.4
oauthlib==3.1.0
opencv-python==4.1.0.25
Pillow==6.0.0
python-crontab==2.3.8
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
virtualenv==16.6.0
zipp==3.1.0
